Add links to different views of the puzzle.
This change adds a views dropdown to the streamer view
navbar that contains links to the differnet supported
views of the puzzle solve.

The streamer view is an existing view and the one that
shows all of the configuration controls in the nav bar.

The participant view is an existing view and one that
shows a read-only view of the solve as it's happening.

The progress view is a newly created view that shows
the progress of a solve as it's happening but only by
shading the filled in cells.  The values put in those
cells are not shown in this view.