When switching into only correct answer mode clear incorrect cells.
This change makes is so that when switching into only correct answer
mode that any cells that have been filled with incorrect values will
be automatically cleared.  Without this it could be confusing to the
streamer when the puzzle is completely filled in, but not solved
becuase a lingering incorrect value hung around from before the
puzzle was in only correct answer mode.