Create initial application structure.
This commit introduces the basic application structure as well as a
mechanism to run the application in development or production mode
in a Docker container.  When in development mode the application is
source file aware and will automatically recompile and relaunch
itself when a source file changes.  In addition while in
development mode template files are served out of the filesystem
instead of from within the application binary.  This allows
templates to be changed and reloaded automatically as well.