Expand the UI to connect to an event stream and have nav menus.
Expands the navbar to have the start/pause button, the views
menu, the settings menu and the puzzle menu.  Only the views
and settings menus are currently wired up.

When on the channel page the UI will connect to the API's event
stream for the channel so that events can be received about
changes to the channel's state (settings or puzzles).