Add bot framework.
The bot framework is comprised of integrations and clients.  An
integration encapsulates the rules of the game being played.
Integrations get to see all messages from a client and can use
them to update its state and call into the API to cause changes
within the application.

A client is an integration with a chat service to receive and
send (currently not supported) messages.

The standard client and one that is used in production is the Twitch
chat integration.  It will join a streamer's channel on twitch.tv and
process puzzle solving commands from the channel.

In addition there is a local client implementation that listens on a
local TCP port and accepts commands entered there.  This client is
used for development purposes to make it possible to create an
integration without having to constantly connect and disconnect from
an external service.