Migrate to use chi instead of gin gonic.
Gin gonic uses httprouter which is very restrictive when it comes to overlapping
URL paths.  This in turn promotes poor REST API design which is undesirable.  As
a result this commit changes things over to use chi
(https://github.com/go-chi/chi) instead.  Chi is very similar to gin gonic in
most ways so the change is quite straightforward.  The only thing missing from
chi is a SSE implementation (there is a partial one in the render package but
it has hardcoded strings for certain error cases which is undesirable.  As a
result a simple, barebones SSE implementation is included in this commit.