Add puz_file_url method for loadig .puz files from a URL.
This is necessary because using the browser to load a URL comes with
quite a few CORS based restrictions.  When loading a URL you can't
change protocols from https to http, and unfortunately most of the
websites that host .puz files seem to not be modern sites with HTTPS
support.  Loading the URL from the server side doesn't have any of
these restrictions however, so it's just easier to ship the URL to
the server and let it load it (which we were already doing for WSJ
puzzles).