Refactor several commonly used types into a model package.
This allows commonly used types (Duration, Status, Font Size) to
be reused across multiple types of puzzles.

In addition this change also refactors the Status type to be a
bit more specific in distinguishing the states a channel can be
in before a puzzle is started.  It now disambiguates between
a channel existing with no puzzle selected and a puzzle being
selected but not yet started.