Add a message router component to the bot.
The message router is used to support multiple integrations
being active at a time across multiple channels.  It receives
notifications of which integrations want to monitor which channels
and then when messages are received from those channels it
routes them to the appropriate message handler(s) of the
integrations that are active in the channel.