Simplify event publishing and make /api/channels reactive to events.
The primary update in this change is to make /api/channels reactive to
events so that connected clients (such as the bot) get immediate
updates when channel states change.  This allows the bot to join the
proper channels much more quickly than having to wait for a periodic
poll to happen to notice the update.

There is still a polling component to the /api/channels route because
there are still operations that happen behind the scenes in the
database like a TTL causing a key to expire that we want clients to
be able to see and react to.