Distinguish between shaded cells and cells with circles.
For crosswords loaded from xwordinfo.com the code will now
pay attention to the `shadecircles` property in the JSON
response and use that to determine whether cells should be
shaded or contain circles.  Unfortunately having both some
circled cells and some shaded cells isn't supported in the
JSON format.

For crosswords loaded from .puz files, shaded cells aren't
supported (because the .puz file doesn't support it) so
only circles are supported.

When rendering, if a circle needs to appear in a box that
also contains a number then an arc is drawn instead of a
full circle so as to not cover up the number at all.

Fixes #3.