Change opacity to be represented by a decimal instead of percentage.
According to https://developer.mozilla.org/en-US/docs/Web/CSS/opacity
support for opacity as a percentage isn't supported by all browsers
yet.  As a result it appears that when webpack minimizes the CSS it
converts the percentage to a decimal value but leaves the percent
symbol!?  So an opacity of `100%` is minimized to `1%` which is
clearly incorrect.

Fixes #6.