Fix small bug in how states are generated for crossword tests.
Previously the NewState method had a bug in it where the cells
field would only be properly generated for square crosswords
because the rows and colums were inadvertently swapped.  This
bug was never noticed because all of the test cases that use
the NewState method are indeed only working with square puzzles.