Render the clues for acrostic puzzles.
This change updates the view to render acrostic clues in 3 columns
that are each independently scrollable.  This should maximize the
full whitespace underneath the puzzle grid.  The font size of the
clue text is controllable through a setting although the font size
of the filled in clue letter and cell number is not currently
controllable.

Closes #26