Update timer component to do less work per update.
Previously the timer would re-parse the total solve duration and the
last start time on every duration update.  The current version only
does parsing when the values being parsed change and during each
update it only computes the elapsed duration.  This will hopefully
make things a little more performant when it comes to updating the
duration.