Fix bug where the genius banner would show on page refresh.
This was happening because detection of crossing the genius
threshold was happening completely in the client based on
the first time it saw the score cross the threshold.  The
problem is when the page is loaded the initial state event
received crosses the threshold so the animation plays.

To fix this the event is now computed on the server side
based on the server observing the score crossing the
threshold when an answer is applied.  This should cause
it to only ever play a single time.

Fixes #19