Normalize available date fetching as well as names for selecting puzzles.
This change centralizes the code for fetching and parsing the available
dates for a particular puzzle type allowing that code to be removed from
each individual puzzle application and put instead in a common place.
In addition it ensures that whenever possible the dates response is
compressed since they can get fairly large over time.

This change also normalizes the names that are used to select specific
puzzle dates across all sources.  They new naming always uses the name
of the puzzle's publication instead of the name of the source where the
puzzle is gotten from.