Disable message processing for some commands of non-solving puzzles.
When channels have a puzzle that's been paused, complete or not yet
started the message handlers shouldn't call the API to modify them.

Doing this cuts down drastically on situations where a command is
being sent to multiple API endpoints because the streamer
previously finished a crossword and is now working on an acrostic.

Closes #38