Add controller application.
The controller application is intended to be used to automate a
channel's displaying of puzzles.  For example it can be used to
switch to a new puzzle after the current one has been solved.

Currently the controller has very minimal functionality it only
works for a set of hard coded channels, only for NYT crosswords,
and only goes back one day.  In the future it will be expanded
to be more versatile.